pyproject.toml
[project]
name = "interview-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "fastapi>=0.115.13",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
]
